backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
numpy
scikit-learn
supabase
google-generativeai
python-dotenv
pydantic
pydantic-settings
